+++ packages/samples/stencil/src/components.d.ts
@@ -10,6 +10,8 @@ import { Router } from "stencil-router-v2";
 export namespace Components {
     interface AppRoot {
     }
+    interface ResultsManager {
+    }
     interface SampleComponent {
     }
     interface SampleResultComponent {
@@ -27,6 +29,12 @@ declare global {
     var HTMLAppRootElement: {
         prototype: HTMLAppRootElement;
         new (): HTMLAppRootElement;
+    };
+    interface HTMLResultsManagerElement extends Components.ResultsManager, HTMLStencilElement {
+    }
+    var HTMLResultsManagerElement: {
+        prototype: HTMLResultsManagerElement;
+        new (): HTMLResultsManagerElement;
     };
     interface HTMLSampleComponentElement extends Components.SampleComponent, HTMLStencilElement {
     }
@@ -54,6 +62,7 @@ declare global {
     };
     interface HTMLElementTagNameMap {
         "app-root": HTMLAppRootElement;
+        "results-manager": HTMLResultsManagerElement;
         "sample-component": HTMLSampleComponentElement;
         "sample-result-component": HTMLSampleResultComponentElement;
         "search-page": HTMLSearchPageElement;
@@ -62,6 +71,8 @@ declare global {
 }
 declare namespace LocalJSX {
     interface AppRoot {
+    }
+    interface ResultsManager {
     }
     interface SampleComponent {
     }
@@ -75,6 +86,7 @@ declare namespace LocalJSX {
     }
     interface IntrinsicElements {
         "app-root": AppRoot;
+        "results-manager": ResultsManager;
         "sample-component": SampleComponent;
         "sample-result-component": SampleResultComponent;
         "search-page": SearchPage;
@@ -86,6 +98,7 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "app-root": LocalJSX.AppRoot & JSXBase.HTMLAttributes<HTMLAppRootElement>;
+            "results-manager": LocalJSX.ResultsManager & JSXBase.HTMLAttributes<HTMLResultsManagerElement>;
             "sample-component": LocalJSX.SampleComponent & JSXBase.HTMLAttributes<HTMLSampleComponentElement>;
             "sample-result-component": LocalJSX.SampleResultComponent & JSXBase.HTMLAttributes<HTMLSampleResultComponentElement>;
             "search-page": LocalJSX.SearchPage & JSXBase.HTMLAttributes<HTMLSearchPageElement>;