+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.37](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.36...@coveo/atomic-angular@2.3.37) (2023-03-10)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.3.36](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.35...@coveo/atomic-angular@2.3.36) (2023-03-02)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.3.35](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.34...@coveo/atomic-angular@2.3.35) (2023-03-02)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.3.34](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.33...@coveo/atomic-angular@2.3.34) (2023-02-28)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
 ## [2.3.33](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.32...@coveo/atomic-angular@2.3.33) (2023-02-22)
 
 **Note:** Version bump only for package @coveo/atomic-angular