+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.36](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.35...@coveo/atomic-react@2.1.36) (2023-03-10)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+## [2.1.35](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.34...@coveo/atomic-react@2.1.35) (2023-03-02)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+## [2.1.34](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.33...@coveo/atomic-react@2.1.34) (2023-03-02)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+## [2.1.33](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.32...@coveo/atomic-react@2.1.33) (2023-02-28)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
 ## [2.1.32](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.31...@coveo/atomic-react@2.1.32) (2023-02-22)
 
 **Note:** Version bump only for package @coveo/atomic-react